extract methods handling the elements array to a separate trait to reuse
