+++ src/FeedIo/Feed/Node.php
@@ -13,17 +13,9 @@ namespace FeedIo\Feed;
 use FeedIo\Feed\Node\Category;
 use FeedIo\Feed\Node\CategoryInterface;
 
-use FeedIo\Feed\Node\Element;
-use FeedIo\Feed\Node\ElementIterator;
-use FeedIo\Feed\Node\ElementInterface;
-
 class Node implements NodeInterface
 {
-
-    /**
-     * @var \ArrayIterator
-     */
-    protected $elements;
+    use ElementsAwareTrait;
     
     /**
      * @var \ArrayIterator
@@ -119,90 +111,6 @@ class Node implements NodeInterface
     public function newCategory()
     {
         return new Category();
-    }
-    
-    /**
-     * @return ElementInterface
-     */
-    public function newElement()
-    {
-        return new Element();
-    }
-
-    /**
-     * @param  string $name element name
-     * @return mixed
-     */
-    public function getValue($name)
-    {
-        foreach ($this->getElementIterator($name) as $element) {
-            return $element->getValue();
-        }
-
-        return;
-    }
-
-    /**
-     * @param  string          $name element name
-     * @return ElementIterator
-     */
-    public function getElementIterator($name)
-    {
-        return new ElementIterator($this->elements, $name);
-    }
-
-    /**
-     * @param  string  $name element name
-     * @return boolean true if the element exists
-     */
-    public function hasElement($name)
-    {
-        $filter = $this->getElementIterator($name);
-
-        return $filter->count() > 0;
-    }
-
-    /**
-     * @param  ElementInterface $element
-     * @return $this
-     */
-    public function addElement(ElementInterface $element)
-    {
-        $this->elements->append($element);
-
-        return $this;
-    }
-
-    /**
-     * Returns all the item's optional elements
-     * @return \ArrayIterator
-     */
-    public function getAllElements()
-    {
-        return $this->elements;
-    }
-
-    /**
-     * Returns the item's optional elements tag names
-     * @return array
-     */
-    public function listElements()
-    {
-        foreach ($this->elements as $element) {
-            yield ($element->getName());
-        }
-    }
-
-    /**
-     * @return \Generator
-     */
-    public function getElementsGenerator()
-    {
-        $elements = $this->getAllElements();
-
-        foreach( $elements as $element ) {
-            yield $element->getName() => $element->getValue();
-        }
     }
 
     /**
